+++ packages/f-header/test/specs/f-header.component.spec.js
@@ -4,7 +4,7 @@ describe('f-header component tests', () => {
     it('should display the f-header component', () => {
         browser.url('http://localhost:8080');
 
-        HeaderComponent.logoIsDisplayed();
+        HeaderComponent.isLogoDisplayed();
     });
 });